+++ screens/HomeScreen.tsx
@@ -1,42 +1,17 @@
-import {StyleSheetI} from "../components/Stylesheet"
-import {Button, View} from "react-native";
-import {ThemeContext} from "../components/ThemeContext";
-import {useContext, useState} from "react";
+import {Discovery} from "./Discovery";
+import {createBottomTabNavigator} from "@react-navigation/bottom-tabs";
+import {ContextExample} from "./ContextExample";
+import {MyRentals} from "./MyRentals";
 
-export function HomeScreen({ navigation } : any) {
-    const theme = useContext(ThemeContext).theme
+const Tab = createBottomTabNavigator();
 
-    // Stylesheet used is the interface from ColorPalette.tsx
-    const styles: StyleSheetI = {
-        container: {
-            flex: 1,
-            backgroundColor: theme.backgroundColor,
-            color: theme.textColor,
-            alignItems: 'center',
-            justifyContent: 'center',
-        },
-        text: {
-            color: theme.textColor,
-            backgroundColor: theme.backgroundColor,
-        },
-        button: {
-            backgroundColor: theme.contrastColor,
-            color: theme.textColor,
-        }
-    }
-
+export function HomeScreen() {
     return (
-        <View style={styles.container}>
-            <Button
-                title="Go to Details"
-                onPress={() => {
-                    /* 1. Navigate to the Details route with params */
-                    navigation.navigate('CarDetailScreen', {
-                        itemId: 86,
-                        otherParam: 'anything you want here',
-                    });
-                }}
-            />
-        </View>
+        <Tab.Navigator>
+            <Tab.Screen name="Discovery" component={Discovery} options={{headerShown: false}} />
+            <Tab.Screen name="MyRentals" component={MyRentals} options={{headerShown: false}} />
+            <Tab.Screen name="Profile" component={ContextExample} options={{headerShown: false}} />
+        </Tab.Navigator>
     );
 }
+
